Project 1 [MV Box]/backend/requirements.txt
python-dotenv~=1.0.1
pymongo~=4.7.2
httpx~=0.27.0
colorama~=0.4.6
fastapi~=0.111.0
pydantic~=2.7.2
uvicorn~=0.30.0
uuid6~=2024.1.12
pytest~=8.2.1
mangum~=0.17.0